third_party/blink/web_tests/fast/dom/Window/closure-access-after-navigation-window-expected.txt

Child 1 loaded.
Child 1 objects set.
getObjectOfChild1().myVal = child1
getGlobalOfChild1().myGlobalVal = child1
Child 2 loaded.
getObjectOfChild1().myVal = child1
getGlobalOfChild1().myGlobalVal = child1
Child 2 objects set.
getObjectOfChild1().myVal = child1
getGlobalOfChild1().myGlobalVal = child1
getObjectOfChild2().myVal = child2
getGlobalOfChild2().myGlobalVal = child2
getObjectOfParent().myVal = parent
getGlobalOfParent().myGlobalVal = parent
